ckan-docker/src/ckanext-regx/requirements.txt
### Ckan API Request Dependencies #####
requests
psycopg2-binary
ckanapi
schedule
python-dotenv

##### Crawler Dependencies #####
requests
beautifulsoup4
google-api-python-client
psycopg2
ckanapi
requests